refactor: renamed ais folder to ai_implementations
- also automatically reformatted some code
- ignore PyCharm files
