+++ mmvb_backend/ai_implementations/api/serializers.py
@@ -1,0 +1,9 @@
+from rest_framework.serializers import ModelSerializer
+
+from ai_implementations.models import AIImplementation
+
+
+class AIImplementationSerializer(ModelSerializer):
+    class Meta:
+        model = AIImplementation
+        fields = ["id", "name", "base_url"]
