+++ mmvb_backend/ai_implementations/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class AisConfig(AppConfig):
-    name = "ais"
+    name = "ai_implementations"
