+++ mmvb_backend/cases/models.py
@@ -16,10 +16,9 @@ class CaseSet(FlowableModel):
 
 
 class Case(FlowableModel, Model):
-    name = models.CharField(unique=True, max_length=150)
     case_sets = models.ManyToManyField("CaseSet", related_name="cases")
     data = JSONField(default=default_data)
     # TODO: relate the case to the company which created it
 
     def __str__(self):
-        return f"{self.name} <{self.status}>"
+        return f"{self.id} <{self.status}>"
