+++ mmvb_backend/urls.py
@@ -15,25 +15,25 @@ Including another URLconf
 """
 from rest_framework.schemas import get_schema_view
 
-from ais.api.urls import router as ais_router
+from ai_implementations.api.urls import router as ai_implementations_router
 from cases.api.urls import router as cases_router
 from common.routers import DefaultRouter
 from django.contrib import admin
 from django.urls import include, path
 from django.views.generic import TemplateView
+from benchmarking_sessions.api.urls import router as benchmarking_sessions_router
 
 router = DefaultRouter(trailing_slash=False)
-router.extend(ais_router)
+router.extend(ai_implementations_router)
 router.extend(cases_router)
+router.extend(benchmarking_sessions_router)
 
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("api/v1/", include(router.urls)),
     path(
         "openapi/",
-        get_schema_view(
-            title="WHO-ITU AI Benchmarking", description="", version="v1"
-        ),
+        get_schema_view(title="WHO-ITU AI Benchmarking", description="", version="v1"),
         name="openapi-schema",
     ),
     path(
